Fix local

+++ finetuning_repo/trl_finetune.py
@@ -355,21 +355,23 @@ def main(args : DictConfig):
 
 
 if __name__ == "__main__":
-    # control cluster wait time
-    with open("./date.txt", "r") as f:
-        d = f.read().strip()
-    queue_time = datetime.strptime(d, '%m/%d/%y %H:%M:%S')
-    launch_time = datetime.now()
-    wait_hours = (launch_time - queue_time).total_seconds() / 3600
     
-    with open("stc/config.yaml", "r") as f:
-        cfg = f.read()   
-         
-    cfg = re.sub("queue_time: .*\n", f"queue_time: {queue_time.strftime('%m/%d/%y %H:%M:%S')}\n", cfg)
-    cfg = re.sub("launch_time: .*\n", f"launch_time: {launch_time.strftime('%m/%d/%y %H:%M:%S')}\n", cfg)
-    cfg = re.sub("wait_hours: .*\n", f"wait_hours: {wait_hours}\n", cfg)
-    
-    with open("stc/config.yaml", "w") as f:
-        f.write(cfg)
+    if os.path.exists("./date.txt"):
+        # control cluster wait time
+        with open("./date.txt", "r") as f:
+            d = f.read().strip()
+        queue_time = datetime.strptime(d, '%m/%d/%y %H:%M:%S')
+        launch_time = datetime.now()
+        wait_hours = (launch_time - queue_time).total_seconds() / 3600
+        
+        with open("stc/config.yaml", "r") as f:
+            cfg = f.read()   
+            
+        cfg = re.sub("queue_time: .*\n", f"queue_time: {queue_time.strftime('%m/%d/%y %H:%M:%S')}\n", cfg)
+        cfg = re.sub("launch_time: .*\n", f"launch_time: {launch_time.strftime('%m/%d/%y %H:%M:%S')}\n", cfg)
+        cfg = re.sub("wait_hours: .*\n", f"wait_hours: {wait_hours}\n", cfg)
+        
+        with open("stc/config.yaml", "w") as f:
+            f.write(cfg)
         
     main()
